Empty trash more often for optimization of WP
EMPTY_TRASH_DAYS gets defined at wp-settings.php, via default-constants.php. So, you have to define it in wp-config.php, if you want to override the default.

Trash can take up a lot of unnecessary room in your website’s database. The bigger the database, the longer it is to retrieve information from it.